move spiffe id validation to registration handler package
