+++ pkg/server/endpoints/registration/handler.go
@@ -6,10 +6,9 @@ import (
 	"net/url"
 	"time"
 
-	uuid "github.com/satori/go.uuid"
+	"github.com/satori/go.uuid"
 	"github.com/sirupsen/logrus"
 	"github.com/spiffe/spire/pkg/common/selector"
-	"github.com/spiffe/spire/pkg/common/util"
 	"github.com/spiffe/spire/pkg/server/catalog"
 	"github.com/spiffe/spire/proto/api/registration"
 	"github.com/spiffe/spire/proto/common"
@@ -32,7 +31,7 @@ func (h *Handler) CreateEntry(
 	response *registration.RegistrationEntryID, err error) {
 
 	// Validate Spiffe ID
-	err = util.ValidateSpiffeID(request.SpiffeId, h.TrustDomain)
+	err = ValidateSpiffeID(request.SpiffeId, h.TrustDomain)
 	if err != nil {
 		h.Log.Error(err)
 		return response, errors.New("Error while validating provided Spiffe ID")
