+++ pkg/server/endpoints/registration/validation_test.go
@@ -1,4 +1,4 @@
-package util
+package registration
 
 import (
 	"net/url"
